finalize moving building system.
