--- src/Sfx-Standalone/.build/tasks/windows/pack.js
@@ -1,7 +1,0 @@
-//-----------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-// Licensed under the MIT License. See License file under the project root for license information.
-//-----------------------------------------------------------------------------
-
-"use strict";
-
