+++ src/Sfx-Standalone/.build/versioning.js
@@ -36,58 +36,59 @@ const buildInfos = config.buildInfos;
  * @param {string} platform 
  * @param {getPackageInfo} getPackageInfo 
  */
-export function generateVersionInfo(platform, getPackageInfo) {
-    if (!utils.isFunction(getPackageInfo)) {
-        throw new Error("getPackageInfo must be supplied.");
-    }
-
-    if (!utils.isObject(buildInfos.updateInfos) || !utils.isString(buildInfos.updateInfos.baseUrl)) {
-        throw new Error("buildInfos.updateInfos.baseUrl must be specified.");
-    }
-
-    /** @type {string} */
-    let channel = "public";
-
-    let prerelease = semver.prerelease(buildInfos.buildNumber);
-
-    if (Array.isArray(prerelease) && prerelease.length > 0) {
-        channel = prerelease[0];
-    }
-
-    /** @type {IVersionInfo} */
-    const versionInfo = {
-        version: buildInfos.buildNumber
-    };
-
-    const baseUrl = String.format("{}/{}/{}", buildInfos.updateInfos.baseUrl, channel, platform);
-    let buildPackageInfo = null;
-
-    if (isObject(buildInfos.updateInfos.packageInfos)) {
-        buildPackageInfo = buildInfos.updateInfos.packageInfos[platform];
-    } else if (!isNullOrUndefined(buildInfos.updateInfos.packageInfos)) {
-        throw new Error("Invalid value for parameter: buildInfos.updateInfos.packageInfos");
-    }
-
-    if (isString(buildPackageInfo)) {
-        versionInfo[platform] = buildPackageInfo;
-    } else if (isNullOrUndefined(buildPackageInfo) || isObject(buildPackageInfo)) {
-        versionInfo[platform] = {};
-
-        for (const arch of buildInfos.targets[platform].archs) {
-            if (isNullOrUndefined(buildPackageInfo) || isNullOrUndefined(buildPackageInfo[arch])) {
-                versionInfo[platform][arch] = getPackageInfo(baseUrl, arch);
-            } else if (isString(buildPackageInfo[arch])) {
-                versionInfo[platform][arch] = utils.format(buildPackageInfo[arch], baseUrl, buildInfos.buildNumber, arch);
-            } else {
-                throw new Error(utils.format("Invalid value for parameter: buildInfos.updateInfos.packageInfos.{}.{}", platform, arch));
-            }
+exports.generateVersionInfo =
+    (platform, getPackageInfo) => {
+        if (!utils.isFunction(getPackageInfo)) {
+            throw new Error("getPackageInfo must be supplied.");
         }
-    } else {
-        throw new Error(utils.format("Invalid value for parameter: buildInfos.updateInfos.packageInfos.{}", platform));
+
+        if (!utils.isObject(buildInfos.updateInfos) || !utils.isString(buildInfos.updateInfos.baseUrl)) {
+            throw new Error("buildInfos.updateInfos.baseUrl must be specified.");
+        }
+
+        /** @type {string} */
+        let channel = "public";
+
+        const prerelease = semver.prerelease(buildInfos.buildNumber);
+
+        if (Array.isArray(prerelease) && prerelease.length > 0) {
+            channel = prerelease[0];
+        }
+
+        /** @type {IVersionInfo} */
+        const versionInfo = {
+            version: buildInfos.buildNumber
+        };
+
+        const baseUrl = String.format("{}/{}/{}", buildInfos.updateInfos.baseUrl, channel, platform);
+        let buildPackageInfo = null;
+
+        if (isObject(buildInfos.updateInfos.packageInfos)) {
+            buildPackageInfo = buildInfos.updateInfos.packageInfos[platform];
+        } else if (!isNullOrUndefined(buildInfos.updateInfos.packageInfos)) {
+            throw new Error("Invalid value for parameter: buildInfos.updateInfos.packageInfos");
+        }
+
+        if (isString(buildPackageInfo)) {
+            versionInfo[platform] = buildPackageInfo;
+        } else if (isNullOrUndefined(buildPackageInfo) || isObject(buildPackageInfo)) {
+            versionInfo[platform] = {};
+
+            for (const arch of buildInfos.targets[platform].archs) {
+                if (isNullOrUndefined(buildPackageInfo) || isNullOrUndefined(buildPackageInfo[arch])) {
+                    versionInfo[platform][arch] = getPackageInfo(baseUrl, arch);
+                } else if (isString(buildPackageInfo[arch])) {
+                    versionInfo[platform][arch] = utils.format(buildPackageInfo[arch], baseUrl, buildInfos.buildNumber, arch);
+                } else {
+                    throw new Error(utils.format("Invalid value for parameter: buildInfos.updateInfos.packageInfos.{}.{}", platform, arch));
+                }
+            }
+        } else {
+            throw new Error(utils.format("Invalid value for parameter: buildInfos.updateInfos.packageInfos.{}", platform));
+        }
+
+        const versionInfoPath = path.resolve(path.join(buildInfos.paths.publishDir, utils.format("version.{}.json", platform)));
+
+        common.ensureDirExists(path.dirname(versionInfoPath));
+        fs.writeFileSync(versionInfoPath, JSON.stringify(versionInfo, null, '\t'));
     }
-
-    const versionInfoPath = path.resolve(path.join(buildInfos.paths.publishDir, utils.format("version.{}.json", platform)));
-
-    common.ensureDirExists(path.dirname(versionInfoPath));
-    fs.writeFileSync(versionInfoPath, JSON.stringify(versionInfo, null, '\t'));
-}
